+++ jenkins/CHANGELOG.md
@@ -1,6 +1,7 @@
 0.7
 
 * Version bump for new base image
+* Fix syslog-ng problems for 4.2 version, remove stats_freq option
 
 ---