+++ jenkins/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * Version bump for new base image
 * Fix syslog-ng problems for 4.2 version, remove stats_freq option
+* Add local unbound and consul DNS services
+* Update README with consul DNS info
 
 ---